Improve editor file structure
