+++ src/pages/dashboard/configsTab/list/ConfigPreviewPanel.tsx
